+++ source/MRViewer/MRViewer.h
@@ -58,28 +58,6 @@ struct LaunchParams
     std::shared_ptr<SplashWindow> splashWindow; // if present will show this window while initializing plugins (after menu initialization)
 };
 
-using FilesLoadedCallback = std::function<void(const std::vector<std::shared_ptr<Object>>& objs,const std::string& errors, const std::string& warnings)>;
-
-struct FileLoadOptions
-{
-    /// first part of undo name
-    const char * undoPrefix = "Open ";
-
-    enum class ReplaceMode
-    {
-        ContructionBased, ///< replace current scene if new one was loaded from single scene file
-        ForceReplace,
-        ForceAdd
-    };
-
-    /// Determines how to deal with current scene after loading new one
-    ReplaceMode replaceMode = ReplaceMode::ContructionBased;
-
-    /// if this callback is set - it is called once when all objects are added to scene
-    /// top level objects only are present here
-    FilesLoadedCallback loadedCallback;
-};
-
 // GLFW-based mesh viewer
 class MRVIEWER_CLASS Viewer
 {
@@ -131,7 +109,8 @@ public:
 
     // Load objects / scenes from files
     // Note! load files with progress bar in next frame if it possible, otherwise load directly inside this function
-    MRVIEWER_API bool loadFiles( const std::vector< std::filesystem::path>& filesList, const FileLoadOptions & options = {} );
+    MRVIEWER_API bool loadFiles( const std::vector< std::filesystem::path>& filesList, const FileLoadOptions & options );
+    MRVIEWER_API bool loadFiles( const std::vector< std::filesystem::path>& filesList );
 
     // Save first selected objects to file
     MRVIEWER_API bool saveToFile( const std::filesystem::path & mesh_file_name );
